[PR python/29603] Disable out-of-scope watchpoints
Currently, when a local software watchpoint goes out of scope, GDB sets the
watchpoint's disposition to `delete at next stop' and then normal stops
(i.e., stop and wait for the next GDB command). When GDB normal stops, it
automatically deletes the breakpoints with their disposition set to
`delete at next stop'.

Suppose a Python script decides not to normal stop when a local software
watchpoint goes out of scope, the watchpoint will not be automatically
deleted even when its disposition is set to `delete at next stop'.

Since GDB single-steps the program and tests the watched expression after each
instruction, not deleting the watchpoint causes the watchpoint to be hit many
more times than it should, as reported in PR python/29603.

This was happening because the watchpoint is not deleted or disabled when
going out of scope.

This commit fixes this issue by disabling the watchpoint when going out of
scope. It also adds a test to ensure this feature isn't regressed in the
future.

Two other solutions seem to solve this issue, but are in fact inappropriate:
1. Automatically delete breakpoints on all kinds of stops
   (in `fetch_inferior_event'). This solution is very slow since the deletion
   requires O(N) time for N breakpoints.
2. Disable the watchpoint after the watchpoint's disposition is set to
   `delete at next stop' (in `watchpoint_del_at_next_stop'). This solution
   modifies a non-extension-related code path, and isn't preferred since this
   issue cannot occur without extension languages. (gdb scripts always normal
   stop before continuing execution)

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=29603